Add Rails app with uploading functionality
This commit adds a Rails app to the repo that satisfies the requirements found at:

https://github.com/lschallenges/data-engineering

These requirements consist of three items:

1. Create an app that recieves a tab delimeted file by way of a form
2. Parse, normalize and store the data from the file in a relational database
3. Display total amount gross revenue to user after the uploaded file has been parsed

The README provides instructions on how to set up and run the application, along with links to the tests that satisfy the requirements.

Thanks for the opportunity to work on this!